refactor: extract validation from contractGovernor.js
